Revert "Merged npcStory and narratorStory modules into story module."
