+++ src/structure/modules/narratorStory/entity.d.ts
@@ -1,0 +1,7 @@
+import { IStage } from "./types";
+
+export interface INarratorEntity {
+  _id: string;
+  episode: number;
+  stages: IStage[];
+}
