+++ src/structure/modules/npcStory/entity.d.ts
@@ -1,0 +1,11 @@
+export interface ILine {
+  intent: string;
+  line: string;
+}
+
+export interface INpcStoryEntity {
+  _id: string;
+  name: string;
+  npcId: string;
+  lines: ILine[];
+}
