+++ src/structure/modules/npcStory/get/dto.ts
