+++ src/structure/modules/npcStory/get/router.ts
