+++ src/structure/modules/npcStory/get/types.d.ts
