+++ src/structure/modules/npcStory/getIntent/dto.ts
