+++ src/structure/modules/npcStory/getIntent/router.ts
