+++ src/structure/modules/npcStory/getIntent/types.d.ts
