+++ src/structure/reqHandler.ts
@@ -6,8 +6,9 @@ import Fights from './modules/fights/handler.js';
 import Inventory from './modules/inventory/handler.js';
 import Map from './modules/maps/handler.js';
 import Message from './modules/message/handler.js';
+import NarratorStory from './modules/narratorStory/handler.js';
 import Npc from './modules/npc/handler.js';
-import Story from './modules/story/handler.js';
+import NpcStory from './modules/npcStory/handler.js';
 import Party from './modules/party/handler.js';
 import Profile from './modules/profile/handler.js';
 import SingleSkill from './modules/singleSkill/handler.js';
@@ -32,7 +33,8 @@ export default class ReqHandler {
   singleSkill: SingleSkill;
   fights: Fights;
   profile: Profile;
-  story: Story;
+  npcStory: NpcStory;
+  narratorStory: NarratorStory;
   message: Message;
   inventory: Inventory;
   bugReport: BugReport;
@@ -61,7 +63,8 @@ export default class ReqHandler {
     this.singleSkill = new SingleSkill(enums.EServices.Users, action);
     this.fights = new Fights(enums.EServices.Fights, action);
     this.profile = new Profile(enums.EServices.Users, action);
-    this.story = new Story(enums.EServices.Story, action);
+    this.npcStory = new NpcStory(enums.EServices.Story, action);
+    this.narratorStory = new NarratorStory(enums.EServices.Story, action);
     this.message = new Message(enums.EServices.Messages, action);
     this.inventory = new Inventory(enums.EServices.Users, action);
     this.bugReport = new BugReport(enums.EServices.Users, action);
