+++ src/structure/router.ts
@@ -7,10 +7,9 @@ import initHealthRoutes from './modules/health/index.js';
 import initInventoryRoutes from './modules/inventory/index.js';
 import initMapRoutes from './modules/maps/index.js';
 import initMessagesRoutes from './modules/message/index.js';
-// import initNarratorStoryRoutes from './modules/narratorStory/index.js';
-import initStoryRoutes from './modules/story/index.js';
+import initNarratorStoryRoutes from './modules/narratorStory/index.js';
 import initNpcRoutes from './modules/npc/index.js';
-// import initNpcStoryRoutes from './modules/npcStory/index.js';
+import initNpcStoryRoutes from './modules/npcStory/index.js';
 import initOidcRoutes from './modules/oidc/index.js';
 import initPartyRoutes from './modules/party/index.js';
 import initProfileRoutes from './modules/profile/index.js';
@@ -60,7 +59,8 @@ export default class AppRouter {
     initSingleSkillRoutes(this.router);
     initFightsRoutes(this.router);
     initNpcRoutes(this.router);
-    initStoryRoutes(this.router);
+    initNarratorStoryRoutes(this.router);
+    initNpcStoryRoutes(this.router);
     initStatsRoutes(this.router);
     initNpcRoutes(this.router);
     initMapRoutes(this.router);
